+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTXDE764YAGJTHQJFTATXX9J"
+        "01KTXSA2VBHJJN0QC6J07G7C0V"
     )
     r"""The unique identifier of the data source"""