+++ packages/orq-rc/RELEASES.md
@@ -8669,3 +8669,13 @@ Based on:
 - [python v4.11.0-rc.16] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.16] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.16 - packages/orq-rc
+
+## 2026-06-12 11:28:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.777.2 (2.903.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.17] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.17] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.17 - packages/orq-rc